Refactor form transform functions to utils
