+++ app/frontend/src/store/modules/form.js
@@ -1,47 +1,7 @@
 import { getField, updateField } from 'vuex-map-fields';
 
 import { formService, rbacService } from '@/services';
-import { IdentityMode, IdentityProviders } from '@/utils/constants';
-
-// TODO: Consider moving or folding these bidirectional transformation functions somewhere else?
-/**
- * @function generateIdps
- * Converts idps and userType to identity provider objects
- * @param {String[]} idps A string array of identity providers
- * @param {String} userType The type of users
- * @returns {Object[]} An object array of identity providers
- */
-function generateIdps({ idps, userType }) {
-  let identityProviders = [];
-  if (userType === IdentityMode.LOGIN) {
-    identityProviders = identityProviders.concat(idps.map((i) => ({ code: i })));
-  } else if (userType === IdentityProviders.PUBLIC) {
-    identityProviders.push(IdentityProviders.PUBLIC);
-  }
-  return identityProviders;
-}
-
-/**
- * @function parseIdps
- * Converts identity provider objects to idps and userType
- * @param {Object[]} identityProviders An object array of identity providers
- * @returns {Object} An object containing idps and userType
- */
-function parseIdps(identityProviders) {
-  const result = {
-    idps: [],
-    userType: 'team',
-  };
-  if (identityProviders.length) {
-    if (identityProviders[0] === IdentityMode.PUBLIC) {
-      result.userType = IdentityMode.PUBLIC;
-    } else {
-      result.userType = IdentityMode.LOGIN;
-      result.idps = identityProviders.map((ip) => ip.code);
-    }
-  }
-  return result;
-}
+import { generateIdps, parseIdps } from '@/utils/transformUtils';
 
 const genInitialForm = () => ({
   description: '',
